New update

+++ GoobySoft/Tools/Software/Algorithms/Algorithms.h
@@ -4,8 +4,8 @@
 #include <string>
 
 std::string Tools_Software_Algorithms_getISO8601Time();
-int_least64_t Tools_Software_Algorithms_getMicroSeconds();
-int_least64_t Tools_Software_Algorithms_getMilliSeconds();
+long long Tools_Software_Algorithms_getMicroSeconds();
+long long Tools_Software_Algorithms_getMilliSeconds();
 void Tools_Software_Algorithms_extractElementFromCharArray(const char arraySeparatedByZeros[], int indexNumber, char extracted[]);
 float Tools_Software_Algorithms_calibration(float x1, float x2, float y1, float y2, float x);
 void Tools_Software_Algorithms_goobySleep(const uint32_t milliseconds);
